Move modifier_image() to more appropriate module
