+++ exts/siborg.create.human/siborg/create/human/mhcaller.py
@@ -3,8 +3,6 @@ import warnings
 import io
 import makehuman
 from pathlib import Path
-import inspect
-import os
 
 # Makehuman loads most modules by manipulating the system path, so we have to
 # run this before we can run the rest of our makehuman imports
@@ -355,30 +353,6 @@ class MHCaller:
             if type in path:
                 return type
         return None
-
-def modifier_image(name : str):
-    """Guess the path to a modifier's corresponding image on disk based on the name
-    of the modifier. Useful for building UI for list of modifiers.
-
-    Parameters
-    ----------
-    name : str
-        Name of the modifier
-
-    Returns
-    -------
-    str
-        The path to the image on disk
-    """
-    if name is None:
-        # If no modifier name is provided, we can't guess the file name
-        return None
-    name = name.lower()
-    # Return the modifier path based on the modifier name
-    # TODO determine if images can be loaded from the Makehuman module stored in
-    # site-packages so we don't have to include the data twice
-    return os.path.join(os.path.dirname(inspect.getfile(makehuman)),targets.getTargets().images.get(name, name))
-
 
 # Create an instance of MHCaller when imported
 MHCaller()
